requirements（以这个为准）.txt
absl-py==2.1.0
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
brotlipy==0.7.0
camelot-py==0.11.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.8
comm==0.2.2
conda==22.11.1
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
contourpy==1.2.1
cpm-kernels==1.0.11
cryptography @ file:///croot/cryptography_1665612644927/work
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.14.6
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
et_xmlfile==2.0.0
exceptiongroup==1.2.1
executing==2.0.1
fastbm25==0.0.2
fastjsonschema==2.19.1
filelock==3.14.0
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.10.0
ghostscript==0.7
greenlet==3.1.1
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.29.3
idna @ file:///croot/idna_1666125576474/work
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab-language-pack-zh-CN==4.2.post1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langchain==0.0.247
langsmith==0.0.92
loguru==0.7.0
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.14
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.4
openapi-schema-pydantic==1.2.4
opencv-python==4.11.0.86
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.1
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.21
Pygments==2.18.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.1.2
pypdf==5.4.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2025.2
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge-chinese==1.0.3
rpds-py==0.18.1
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.5.3
Send2Trash==1.8.3
sentencepiece==0.1.99
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.40
stack-data==0.6.3
supervisor==4.2.5
sympy==1.12.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
terminado==0.18.1
tinycss2==1.3.0
tokenizers==0.13.3
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch @ http://download.pytorch.org/whl/cu118/torch-2.1.2%2Bcu118-cp310-cp310-linux_x86_64.whl
torchvision @ http://download.pytorch.org/whl/cu118/torchvision-0.16.2%2Bcu118-cp310-cp310-linux_x86_64.whl
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.30.2
triton==2.1.0
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1670526988650/work
Wand==0.6.13
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
xxhash==3.5.0
yarl==1.18.3

# --- 政策法规 FAISS RAG（可选，见 docs/policy_vector_retrieval.py build）---
sentence-transformers>=2.2.0
faiss-cpu>=1.8

# --- 用户管理 / JWT / 审计（MySQL 与 MYSQL_CONFIG 同实例，表 qa_*）---
sqlalchemy>=2.0
pymysql>=1.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
